Renamed redis-cached-script.
