+++ src/persistence/primary/script/update-entities-cached-script-set.ts
@@ -1,7 +1,7 @@
 import { CacheMode } from '../options/cache-mode';
 import { PersistencyUpdateOptions } from '../options/persistency-update-options';
 import { IRedisCachedScriptSet } from './IRedisCachedScriptSet';
-import { RedisCachedScript } from './RedisCachedScript';
+import { RedisCachedScript } from './redis-cached-script';
 
 export class UpdateEntitiesCachedScriptSet implements IRedisCachedScriptSet<PersistencyUpdateOptions> {
   /**
